refactor: decompose progress_db into db/ domain modules (#282)
