+++ src/pyimgtag/progress_db.py
@@ -1,1835 +1,24 @@
-"""SQLite progress database for incremental image processing."""
+"""Compatibility re-export for the progress database (moved to :mod:`pyimgtag.db`).
+
+The SQLite progress database was decomposed into the ``pyimgtag.db``
+package (issue #282): :class:`~pyimgtag.db.progress_db.ProgressDB` is now a
+thin facade over the ``ImageDB`` / ``FaceDB`` / ``JudgeDB`` domain helpers.
+This module keeps the historical ``pyimgtag.progress_db`` import path
+working unchanged for all existing callers.
+"""
 
 from __future__ import annotations
 
-import json
-import logging
-import sqlite3
-import struct
-from collections import defaultdict
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import TYPE_CHECKING, Any
-
+from pyimgtag.db.image_db import _DEFAULT_PATH_BATCH_SIZE
+from pyimgtag.db.judge_db import _DEFAULT_JUDGE_RESULTS_LIMIT
+from pyimgtag.db.progress_db import ProgressDB
 from pyimgtag.models import FaceDetection, ImageResult, PersonCluster
 
-if TYPE_CHECKING:
-    from collections.abc import Iterator
-
-    import numpy as np
-
-    from pyimgtag.models import JudgeResult
-
-logger = logging.getLogger(__name__)
-
-# Default limit for get_all_judge_results; kept as a named constant so it
-# is easy to find and change in one place.
-_DEFAULT_JUDGE_RESULTS_LIMIT: int = 200
-
-# Default batch size for iter_image_paths; 1 000 rows per round-trip
-# keeps memory usage low on a 22 k-photo library.
-_DEFAULT_PATH_BATCH_SIZE: int = 1000
-
-
-class ProgressDB:
-    """Track which images have been processed to enable incremental re-runs."""
-
-    # Each entry is (target_version, sql_statement).
-    # Version 1 is the baseline schema created by _create_table().
-    # Migrations are applied in ascending version order on every open.
-    _MIGRATIONS: tuple[tuple[int, str], ...] = (
-        (2, "ALTER TABLE processed_images ADD COLUMN scene_category TEXT"),
-        (2, "ALTER TABLE processed_images ADD COLUMN emotional_tone TEXT"),
-        (2, "ALTER TABLE processed_images ADD COLUMN cleanup_class TEXT"),
-        (2, "ALTER TABLE processed_images ADD COLUMN has_text INTEGER DEFAULT 0"),
-        (2, "ALTER TABLE processed_images ADD COLUMN text_summary TEXT"),
-        (2, "ALTER TABLE processed_images ADD COLUMN event_hint TEXT"),
-        (2, "ALTER TABLE processed_images ADD COLUMN significance TEXT"),
-        (
-            3,
-            """CREATE TABLE IF NOT EXISTS persons (
-                id        INTEGER PRIMARY KEY AUTOINCREMENT,
-                label     TEXT NOT NULL DEFAULT '',
-                confirmed INTEGER NOT NULL DEFAULT 0
-            )""",
-        ),
-        (
-            3,
-            """CREATE TABLE IF NOT EXISTS faces (
-                id         INTEGER PRIMARY KEY AUTOINCREMENT,
-                image_path TEXT NOT NULL,
-                bbox_x     INTEGER NOT NULL DEFAULT 0,
-                bbox_y     INTEGER NOT NULL DEFAULT 0,
-                bbox_w     INTEGER NOT NULL DEFAULT 0,
-                bbox_h     INTEGER NOT NULL DEFAULT 0,
-                confidence REAL NOT NULL DEFAULT 0.0,
-                embedding  BLOB,
-                person_id  INTEGER REFERENCES persons(id)
-            )""",
-        ),
-        (3, "CREATE INDEX IF NOT EXISTS idx_faces_image ON faces(image_path)"),
-        (3, "CREATE INDEX IF NOT EXISTS idx_faces_person ON faces(person_id)"),
-        (4, "ALTER TABLE processed_images ADD COLUMN nearest_city TEXT"),
-        (4, "ALTER TABLE processed_images ADD COLUMN nearest_region TEXT"),
-        (4, "ALTER TABLE processed_images ADD COLUMN nearest_country TEXT"),
-        (
-            5,
-            """CREATE TABLE IF NOT EXISTS judge_scores (
-                file_path          TEXT PRIMARY KEY,
-                scored_at          TEXT NOT NULL,
-                weighted_score     REAL NOT NULL,
-                core_score         REAL NOT NULL,
-                visible_score      REAL NOT NULL,
-                verdict            TEXT,
-                impact             REAL,
-                story_subject      REAL,
-                composition_center REAL,
-                lighting           REAL,
-                creativity_style   REAL,
-                color_mood         REAL,
-                presentation_crop  REAL,
-                technical_excellence REAL,
-                focus_sharpness    REAL,
-                exposure_tonal     REAL,
-                noise_cleanliness  REAL,
-                subject_separation REAL,
-                edit_integrity     REAL
-            )""",
-        ),
-        (6, "ALTER TABLE persons ADD COLUMN source TEXT NOT NULL DEFAULT 'auto'"),
-        (6, "ALTER TABLE persons ADD COLUMN trusted INTEGER NOT NULL DEFAULT 0"),
-        # 0.10.0: simple-prompt judge stores the model's natural-language
-        # justification next to the integer score.
-        (7, "ALTER TABLE judge_scores ADD COLUMN reason TEXT"),
-        # 0.12.0: persist the photo's EXIF capture timestamp so the Query
-        # page can show "when the photo was taken" without re-reading EXIF.
-        (8, "ALTER TABLE processed_images ADD COLUMN image_date TEXT"),
-        # 0.13.6: indexes on filter/sort columns to speed up paginated queries.
-        (9, "CREATE INDEX IF NOT EXISTS idx_pi_status ON processed_images(status)"),
-        (9, "CREATE INDEX IF NOT EXISTS idx_pi_cleanup ON processed_images(cleanup_class)"),
-        (9, "CREATE INDEX IF NOT EXISTS idx_pi_date ON processed_images(processed_at)"),
-        # 0.16.6: faces can be explicitly dismissed so auto-clustering skips them.
-        (10, "ALTER TABLE faces ADD COLUMN ignored INTEGER NOT NULL DEFAULT 0"),
-        (10, "CREATE INDEX IF NOT EXISTS idx_faces_ignored ON faces(ignored)"),
-        # 0.16.6: track images already face-scanned so zero-face images are not
-        # re-scanned on every subsequent faces scan run.
-        (
-            11,
-            """CREATE TABLE IF NOT EXISTS face_scanned_images (
-                image_path TEXT PRIMARY KEY
-            )""",
-        ),
-    )
-
-    def __init__(self, db_path: str | Path | None = None) -> None:
-        """Open (creating if needed) the SQLite progress database.
-
-        Args:
-            db_path: Path to the database file. When ``None``, defaults to
-                ``~/.cache/pyimgtag/progress.db``.
-
-        The parent directory is created if missing, the connection is opened
-        with ``check_same_thread=False`` (a background thread may read while the
-        main thread writes) in WAL journal mode, and the schema plus any pending
-        versioned migrations run on open.
-
-        Raises:
-            sqlite3.DatabaseError: If the file is not a usable database or a
-                migration cannot be applied.
-        """
-        if db_path is None:
-            db_path = Path.home() / ".cache" / "pyimgtag" / "progress.db"
-        self._path = Path(db_path)
-        self._path.parent.mkdir(parents=True, exist_ok=True)
-        self._conn = sqlite3.connect(str(self._path), check_same_thread=False)
-        self._conn.execute("PRAGMA journal_mode=WAL")
-        self._create_table()
-
-    @property
-    def path(self) -> Path:
-        """Filesystem path to the backing SQLite database file."""
-        return self._path
-
-    def _create_table(self) -> None:
-        self._conn.execute(
-            """
-            CREATE TABLE IF NOT EXISTS processed_images (
-                file_path    TEXT PRIMARY KEY,
-                file_size    INTEGER,
-                file_mtime   REAL,
-                tags         TEXT,
-                scene_summary TEXT,
-                processed_at TEXT,
-                status       TEXT,
-                error_message TEXT
-            )
-            """
-        )
-        self._conn.commit()
-        # Mark a brand-new database (user_version == 0) as version 1 so that
-        # existing migration entries for version > 1 still apply on first open.
-        current: int = self._conn.execute("PRAGMA user_version").fetchone()[0]
-        if current == 0:
-            self._conn.execute("PRAGMA user_version = 1")
-        self._migrate()
-
-    def _migrate(self) -> None:
-        """Apply any pending versioned migrations and update user_version."""
-        current: int = self._conn.execute("PRAGMA user_version").fetchone()[0]
-        # Collect all target versions that are still pending.
-        pending_versions = sorted({ver for ver, _ in self._MIGRATIONS if ver > current})
-        if not pending_versions:
-            return
-        for target_ver in pending_versions:
-            stmts = [sql for ver, sql in self._MIGRATIONS if ver == target_ver]
-            self._conn.execute(f"SAVEPOINT migrate_v{int(target_ver)}")  # nosec B608
-            try:
-                for sql in stmts:
-                    try:
-                        self._conn.execute(sql)
-                    except sqlite3.OperationalError as exc:
-                        # Tolerate "duplicate column name" so re-running migrations
-                        # on an already-migrated DB is safe.
-                        if "duplicate column name" not in str(exc).lower():
-                            raise
-                self._conn.execute(f"PRAGMA user_version = {int(target_ver)}")  # nosec B608
-                self._conn.execute(f"RELEASE migrate_v{int(target_ver)}")  # nosec B608
-            except Exception:  # noqa: BLE001 — roll back savepoint on any failure, then re-raise
-                self._conn.execute(f"ROLLBACK TO migrate_v{int(target_ver)}")  # nosec B608
-                raise
-        self._conn.commit()
-
-    def is_processed(self, file_path: Path) -> bool:
-        """Return True if the file was already successfully processed and hasn't changed.
-
-        Rows with status != 'ok' (e.g. transient Ollama failures) are treated as
-        not processed so they will be retried on the next run.
-        """
-        row = self._conn.execute(
-            "SELECT file_size, file_mtime, status FROM processed_images WHERE file_path = ?",
-            (str(file_path),),
-        ).fetchone()
-        if row is None:
-            return False
-        if row[2] != "ok":
-            return False
-        try:
-            stat = file_path.stat()
-        except OSError:
-            return False
-        return row[0] == stat.st_size and row[1] == stat.st_mtime
-
-    def mark_done(self, file_path: Path, result: ImageResult) -> None:
-        """Record a processed image result.
-
-        If ``file_path.stat()`` fails (permission error, vanished file), the
-        stored size/mtime fall back to ``0``/``0.0``. Such a row can never match
-        a real file in :meth:`is_processed`, so the file would be re-processed on
-        every subsequent run; the failure is logged at debug level.
-        """
-        try:
-            stat = file_path.stat()
-            size = stat.st_size
-            mtime = stat.st_mtime
-        except OSError as e:
-            logger.debug("stat() failed for %s; recording size/mtime as 0: %s", file_path, e)
-            size = 0
-            mtime = 0.0
-        self._conn.execute(
-            """
-            INSERT OR REPLACE INTO processed_images
-                (file_path, file_size, file_mtime, tags, scene_summary,
-                 processed_at, status, error_message,
-                 scene_category, emotional_tone, cleanup_class, has_text,
-                 text_summary, event_hint, significance,
-                 nearest_city, nearest_region, nearest_country, image_date)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-            (
-                str(file_path),
-                size,
-                mtime,
-                json.dumps(result.tags),
-                result.scene_summary,
-                datetime.now(timezone.utc).isoformat(),
-                result.processing_status,
-                result.error_message,
-                result.scene_category,
-                result.emotional_tone,
-                result.cleanup_class,
-                1 if result.has_text else 0,
-                result.text_summary,
-                result.event_hint,
-                result.significance,
-                result.nearest_city,
-                result.nearest_region,
-                result.nearest_country,
-                result.image_date,
-            ),
-        )
-        self._conn.commit()
-
-    def get_stats(self) -> dict:
-        """Return counts of processed images by status."""
-        total = self._conn.execute("SELECT COUNT(*) FROM processed_images").fetchone()[0]
-        ok = self._conn.execute(
-            "SELECT COUNT(*) FROM processed_images WHERE status = 'ok'"
-        ).fetchone()[0]
-        error = self._conn.execute(
-            "SELECT COUNT(*) FROM processed_images WHERE status = 'error'"
-        ).fetchone()[0]
-        return {"total": total, "ok": ok, "error": error}
-
-    def reset_all(self) -> int:
-        """Delete all rows from the processed_images table. Returns count deleted."""
-        cur = self._conn.execute("SELECT COUNT(*) FROM processed_images")
-        count = cur.fetchone()[0]
-        self._conn.execute("DELETE FROM processed_images")
-        self._conn.commit()
-        return count
-
-    def reset_by_status(self, status: str) -> int:
-        """Delete rows with the given status (e.g. 'error'). Returns count deleted."""
-        cur = self._conn.execute(
-            "SELECT COUNT(*) FROM processed_images WHERE status = ?", (status,)
-        )
-        count = cur.fetchone()[0]
-        self._conn.execute("DELETE FROM processed_images WHERE status = ?", (status,))
-        self._conn.commit()
-        return count
-
-    _QUERY_SORTS: dict[str, str] = {
-        "path_asc": "pi.file_path ASC",
-        "path_desc": "pi.file_path DESC",
-        "newest": "pi.processed_at DESC, pi.file_path ASC",
-        "oldest": "pi.processed_at ASC, pi.file_path ASC",
-        "judge_desc": "js.weighted_score DESC NULLS LAST, pi.file_path ASC",
-        "judge_asc": "js.weighted_score ASC NULLS LAST, pi.file_path ASC",
-        "shot_desc": "pi.image_date DESC NULLS LAST, pi.file_path ASC",
-        "shot_asc": "pi.image_date ASC NULLS LAST, pi.file_path ASC",
-    }
-
-    @staticmethod
-    def _build_query_conditions(
-        tag: str | None,
-        has_text: bool | None,
-        cleanup_class: str | None,
-        scene_category: str | None,
-        city: str | None,
-        country: str | None,
-        status: str | None,
-        min_judge_score: int | None,
-        max_judge_score: int | None,
-        judged: bool | None,
-    ) -> tuple[list[str], list[object]]:
-        """Translate filter arguments into ``(conditions, params)`` for query_images."""
-        conditions: list[str] = []
-        params: list[object] = []
-
-        if tag is not None:
-            conditions.append(
-                "EXISTS (SELECT 1 FROM json_each(pi.tags) WHERE LOWER(value) LIKE LOWER(?))"
-            )
-            params.append(f"%{tag}%")
-        if has_text is True:
-            conditions.append("pi.has_text = 1")
-        elif has_text is False:
-            conditions.append("(pi.has_text = 0 OR pi.has_text IS NULL)")
-        if cleanup_class is not None:
-            conditions.append("pi.cleanup_class = ?")
-            params.append(cleanup_class)
-        if scene_category is not None:
-            conditions.append("pi.scene_category = ?")
-            params.append(scene_category)
-        if city is not None:
-            conditions.append("LOWER(pi.nearest_city) LIKE LOWER(?)")
-            params.append(f"%{city}%")
-        if country is not None:
-            conditions.append("LOWER(pi.nearest_country) LIKE LOWER(?)")
-            params.append(f"%{country}%")
-        if status is not None:
-            conditions.append("pi.status = ?")
-            params.append(status)
-        if min_judge_score is not None:
-            conditions.append("js.weighted_score >= ?")
-            params.append(min_judge_score)
-        if max_judge_score is not None:
-            conditions.append("js.weighted_score <= ?")
-            params.append(max_judge_score)
-        if judged is True:
-            conditions.append("js.weighted_score IS NOT NULL")
-        elif judged is False:
-            conditions.append("js.weighted_score IS NULL")
-
-        return conditions, params
-
-    @staticmethod
-    def _build_images_query(conditions: list[str], order_clause: str, limit: int | None) -> str:
-        """Assemble the SELECT SQL for query_images from pre-validated parts."""
-        where = ("WHERE " + " AND ".join(conditions)) if conditions else ""
-        limit_clause = f"LIMIT {int(limit)}" if limit is not None else ""
-        return (  # nosec B608
-            "SELECT pi.file_path, pi.tags, pi.scene_summary, pi.processed_at, pi.status, "
-            "pi.cleanup_class, pi.scene_category, pi.emotional_tone, pi.event_hint, "
-            "pi.significance, pi.nearest_city, pi.nearest_region, pi.nearest_country, "
-            "pi.error_message, js.weighted_score, js.reason, js.verdict, pi.image_date "
-            "FROM processed_images pi "
-            "LEFT JOIN judge_scores js ON js.file_path = pi.file_path "
-            + where  # nosec B608
-            + f" ORDER BY {order_clause} "  # nosec B608
-            + limit_clause
-        )
-
-    def query_images(
-        self,
-        tag: str | None = None,
-        has_text: bool | None = None,
-        cleanup_class: str | None = None,
-        scene_category: str | None = None,
-        city: str | None = None,
-        country: str | None = None,
-        status: str | None = None,
-        limit: int | None = None,
-        min_judge_score: int | None = None,
-        max_judge_score: int | None = None,
-        judged: bool | None = None,
-        sort: str = "path_asc",
-    ) -> list[dict]:
-        """Query images with advanced filters.
-
-        Args:
-            tag: Case-insensitive substring match against any tag value.
-            has_text: True = only images with text; False = only without; None = any.
-            cleanup_class: Exact match against cleanup_class ('delete', 'review', etc.).
-            scene_category: Exact match against scene_category.
-            city: Case-insensitive substring match against nearest_city.
-            country: Case-insensitive substring match against nearest_country.
-            status: Exact match against status ('ok', 'error').
-            limit: Max rows to return. None = no limit.
-            min_judge_score: Only return images whose judge weighted_score >= this value.
-            max_judge_score: Only return images whose judge weighted_score <= this value.
-            judged: True = only images with a judge_scores row; False = only un-judged;
-                None = any.
-            sort: One of ``path_asc`` (default), ``path_desc``, ``newest``,
-                ``oldest``, ``judge_desc``, ``judge_asc``, ``shot_desc``,
-                ``shot_asc``.
-
-        Returns:
-            List of image metadata dicts.
-        """
-        conditions, params = self._build_query_conditions(
-            tag,
-            has_text,
-            cleanup_class,
-            scene_category,
-            city,
-            country,
-            status,
-            min_judge_score,
-            max_judge_score,
-            judged,
-        )
-        order_clause = self._QUERY_SORTS.get(sort, self._QUERY_SORTS["path_asc"])
-        query = self._build_images_query(conditions, order_clause, limit)
-        rows = self._conn.execute(query, params).fetchall()
-        return [self._query_row_to_dict(r) for r in rows]
-
-    @staticmethod
-    def _query_row_to_dict(row: tuple) -> dict:
-        """Convert a query_images SELECT row to a metadata dict.
-
-        Columns 14-16 (``js.weighted_score``, ``js.reason``, ``js.verdict``)
-        come from the LEFT JOIN with ``judge_scores`` and are ``None`` for any
-        image that has not been judged yet; the trailing column 17 is
-        ``pi.image_date``.
-        """
-        file_path: str = row[0]
-        tags_raw: str | None = row[1]
-        try:
-            tags_list: list[str] = json.loads(tags_raw) if tags_raw else []
-        except (json.JSONDecodeError, TypeError):
-            tags_list = []
-        weighted_raw = row[14] if len(row) > 14 else None
-        return {
-            "file_path": file_path,
-            "file_name": Path(file_path).name,
-            "tags_list": tags_list,
-            "scene_summary": row[2],
-            "processed_at": row[3],
-            "status": row[4],
-            "cleanup_class": row[5],
-            "scene_category": row[6],
-            "emotional_tone": row[7],
-            "event_hint": row[8],
-            "significance": row[9],
-            "nearest_city": row[10],
-            "nearest_region": row[11],
-            "nearest_country": row[12],
-            "error_message": row[13] if len(row) > 13 else None,
-            "judge_score": int(round(float(weighted_raw))) if weighted_raw is not None else None,
-            "judge_reason": row[15] if len(row) > 15 else None,
-            "judge_verdict": row[16] if len(row) > 16 else None,
-            "image_date": row[17] if len(row) > 17 else None,
-        }
-
-    def get_tag_counts(self) -> list[tuple[str, int]]:
-        """Return (tag, count) pairs sorted by count descending.
-
-        Counts how many images have each distinct tag.
-
-        Returns:
-            List of (tag_name, image_count) tuples.
-        """
-        rows = self._conn.execute(
-            "SELECT LOWER(value), COUNT(*) AS cnt "
-            "FROM processed_images, json_each(tags) "
-            "WHERE tags IS NOT NULL "
-            "GROUP BY LOWER(value) "
-            "ORDER BY cnt DESC, LOWER(value)"
-        ).fetchall()
-        return [(r[0], r[1]) for r in rows]
-
-    def _iter_tag_rows(self) -> list[tuple[str, list[str]]]:
-        """Return ``(file_path, tags)`` for every row that has tags.
-
-        Used by :meth:`rename_tag`, :meth:`delete_tag`, and
-        :meth:`merge_tags` to share the fetch-parse boilerplate.  Rows
-        whose ``tags`` column cannot be parsed as JSON are silently
-        excluded.
-        """
-        rows = self._conn.execute(
-            "SELECT file_path, tags FROM processed_images WHERE tags IS NOT NULL"
-        ).fetchall()
-        result = []
-        for file_path, tags_raw in rows:
-            try:
-                tags: list[str] = json.loads(tags_raw)
-            except (json.JSONDecodeError, TypeError):
-                continue
-            result.append((file_path, tags))
-        return result
-
-    def _write_tags(self, file_path: str, new_tags: list[str]) -> None:
-        """Overwrite the tags for *file_path* inside an open transaction."""
-        self._conn.execute(
-            "UPDATE processed_images SET tags = ? WHERE file_path = ?",
-            (json.dumps(new_tags), file_path),
-        )
-
-    def rename_tag(self, old_tag: str, new_tag: str) -> int:
-        """Rename a tag across all images.
-
-        Replaces every occurrence of *old_tag* (case-insensitive) with *new_tag*
-        (lowercased). Images that already contain *new_tag* are de-duplicated so
-        the tag only appears once.
-
-        Args:
-            old_tag: The tag to rename.
-            new_tag: The replacement tag (will be lowercased).
-
-        Returns:
-            Number of images updated.
-        """
-        old_lower = old_tag.lower()
-        new_lower = new_tag.lower()
-        updated = 0
-        with self._conn:
-            for file_path, tags in self._iter_tag_rows():
-                tags_lower = [t.lower() for t in tags]
-                if old_lower not in tags_lower:
-                    continue
-                new_tags: list[str] = []
-                seen: set[str] = set()
-                for t in tags:
-                    val = t.lower()
-                    replacement = new_lower if val == old_lower else val
-                    if replacement not in seen:
-                        seen.add(replacement)
-                        new_tags.append(replacement)
-                self._write_tags(file_path, new_tags)
-                updated += 1
-        return updated
-
-    def delete_tag(self, tag: str) -> int:
-        """Remove a tag from all images that have it.
-
-        Args:
-            tag: The tag to delete (case-insensitive match).
-
-        Returns:
-            Number of images updated.
-        """
-        tag_lower = tag.lower()
-        updated = 0
-        with self._conn:
-            for file_path, tags in self._iter_tag_rows():
-                new_tags = [t for t in tags if t.lower() != tag_lower]
-                if len(new_tags) == len(tags):
-                    continue
-                self._write_tags(file_path, new_tags)
-                updated += 1
-        return updated
-
-    def merge_tags(self, source_tag: str, target_tag: str) -> int:
-        """Merge *source_tag* into *target_tag*.
-
-        For every image that has *source_tag*, removes it and adds *target_tag*
-        (if not already present). The *target_tag* is lowercased.
-
-        Args:
-            source_tag: The tag to replace (case-insensitive match).
-            target_tag: The tag to add (will be lowercased).
-
-        Returns:
-            Number of images updated.
-        """
-        src_lower = source_tag.lower()
-        tgt_lower = target_tag.lower()
-        updated = 0
-        with self._conn:
-            for file_path, tags in self._iter_tag_rows():
-                tags_lower = [t.lower() for t in tags]
-                if src_lower not in tags_lower:
-                    continue
-                new_tags = [t for t in tags if t.lower() != src_lower]
-                if tgt_lower not in [t.lower() for t in new_tags]:
-                    new_tags.append(tgt_lower)
-                self._write_tags(file_path, new_tags)
-                updated += 1
-        return updated
-
-    def get_cleanup_candidates(self, include_review: bool = False) -> list[dict]:
-        """Return photos flagged for cleanup.
-
-        Returns list of dicts with keys: file_path, file_name (basename),
-        cleanup_class, tags, scene_summary, image_date, nearest_city, nearest_country.
-        Always includes cleanup_class='delete'. If include_review=True, also includes 'review'.
-        Orders by file_path.
-        """
-        if include_review:
-            placeholders = "?, ?"
-            params: tuple = ("delete", "review")
-        else:
-            placeholders = "?"
-            params = ("delete",)
-        # Parameterized query; placeholders are code-controlled literals
-        query = (  # nosec B608
-            "SELECT file_path, tags, scene_summary, image_date, cleanup_class, "
-            "nearest_city, nearest_country "
-            "FROM processed_images "
-            "WHERE cleanup_class IN ("
-            + placeholders  # nosec B608
-            + ") "
-            "ORDER BY file_path"
-        )
-        rows = self._conn.execute(query, params).fetchall()
-
-        result = []
-        for row in rows:
-            file_path = row[0]
-            result.append(
-                {
-                    "file_path": file_path,
-                    "file_name": Path(file_path).name,
-                    "tags": row[1],
-                    "scene_summary": row[2],
-                    "image_date": row[3],
-                    "cleanup_class": row[4],
-                    "nearest_city": row[5],
-                    "nearest_country": row[6],
-                }
-            )
-        return result
-
-    # --- review UI query/update methods ---
-
-    # SQLite basename idiom over the separator-normalized path NORM
-    # (backslashes folded to '/' so Windows paths work too):
-    # replace(NORM, '/', '') yields the set of non-slash characters,
-    # rtrim(NORM, <that>) strips the basename off the right (stopping at the
-    # last '/'), and the outer replace() removes that directory prefix.
-    # A file_path tie-breaker keeps pagination stable for duplicate basenames.
-    _NORM_PATH = "replace(file_path, '\\', '/')"
-    _NAME_SORT_EXPR = (
-        f"LOWER(replace({_NORM_PATH}, rtrim({_NORM_PATH}, replace({_NORM_PATH}, '/', '')), ''))"
-    )
-
-    _GET_IMAGES_SORTS: dict[str, str] = {
-        "path_asc": "file_path ASC",
-        "path_desc": "file_path DESC",
-        "newest": "processed_at DESC, file_path ASC",
-        "oldest": "processed_at ASC, file_path ASC",
-        "name_asc": f"{_NAME_SORT_EXPR} ASC, file_path ASC",
-        "name_desc": f"{_NAME_SORT_EXPR} DESC, file_path DESC",
-    }
-
-    def get_images(
-        self,
-        limit: int = 50,
-        offset: int = 0,
-        status: str | None = None,
-        cleanup_class: str | None = None,
-        sort: str = "path_asc",
-    ) -> list[dict]:
-        """Return paginated image records with all metadata.
-
-        Args:
-            limit: Max rows to return.
-            offset: Row offset for pagination.
-            status: Optional filter by status ('ok', 'error').
-            cleanup_class: Optional filter by cleanup_class ('delete', 'review').
-            sort: One of ``path_asc`` (default), ``path_desc``, ``newest``,
-                ``oldest``, ``name_asc``, ``name_desc``.
-
-        Returns:
-            List of image metadata dicts.
-        """
-        conditions: list[str] = []
-        params: list[object] = []
-        if status is not None:
-            conditions.append("status = ?")
-            params.append(status)
-        if cleanup_class is not None:
-            conditions.append("cleanup_class = ?")
-            params.append(cleanup_class)
-        order_clause = self._GET_IMAGES_SORTS.get(sort, self._GET_IMAGES_SORTS["path_asc"])
-        # All filter columns live on ``processed_images`` and the join key
-        # ``file_path`` is the only column that collides with judge_scores,
-        # so qualify each predicate with the ``pi.`` alias.
-        joined_where = "WHERE " + " AND ".join(f"pi.{c}" for c in conditions) if conditions else ""
-        # Sort clauses reference unambiguous columns (file_path resolves to
-        # processed_images via the alias, processed_at and the LOWER()
-        # variants only exist on processed_images).
-        order_qualified = order_clause.replace("file_path", "pi.file_path").replace(
-            "processed_at", "pi.processed_at"
-        )
-        query = (  # nosec B608
-            "SELECT pi.file_path, pi.tags, pi.scene_summary, pi.processed_at, "
-            "pi.status, pi.cleanup_class, pi.scene_category, pi.emotional_tone, "
-            "pi.event_hint, pi.significance, pi.error_message, "
-            "js.weighted_score, js.verdict, js.reason "
-            "FROM processed_images pi "
-            "LEFT JOIN judge_scores js ON js.file_path = pi.file_path "
-            + joined_where  # nosec B608
-            + f" ORDER BY {order_qualified} LIMIT ? OFFSET ?"  # nosec B608
-        )
-        params.extend([limit, offset])
-        rows = self._conn.execute(query, params).fetchall()
-        return [self._image_row_to_dict(r) for r in rows]
-
-    def count_images(
-        self,
-        status: str | None = None,
-        cleanup_class: str | None = None,
-    ) -> int:
-        """Count images matching optional filters."""
-        conditions: list[str] = []
-        params: list[object] = []
-        if status is not None:
-            conditions.append("status = ?")
-            params.append(status)
-        if cleanup_class is not None:
-            conditions.append("cleanup_class = ?")
-            params.append(cleanup_class)
-        where = ("WHERE " + " AND ".join(conditions)) if conditions else ""
-        query = "SELECT COUNT(*) FROM processed_images " + where  # nosec B608
-        return self._conn.execute(query, params).fetchone()[0]
-
-    def get_image(self, file_path: str) -> dict | None:
-        """Return metadata for a single image, or None if not found.
-
-        The join with ``judge_scores`` brings the weighted-judge score and
-        verdict back so the review UI and dashboard click-through can show
-        them on the per-image card without a follow-up request.
-        """
-        row = self._conn.execute(
-            "SELECT pi.file_path, pi.tags, pi.scene_summary, pi.processed_at, "
-            "pi.status, pi.cleanup_class, pi.scene_category, pi.emotional_tone, "
-            "pi.event_hint, pi.significance, pi.error_message, "
-            "js.weighted_score, js.verdict, js.reason "
-            "FROM processed_images pi "
-            "LEFT JOIN judge_scores js ON js.file_path = pi.file_path "
-            "WHERE pi.file_path = ?",
-            (file_path,),
-        ).fetchone()
-        if row is None:
-            return None
-        return self._image_row_to_dict(row)
-
-    def get_known_file_path(self, file_path: str) -> str | None:
-        """Return the stored path if known to tagging **or** judging, else None.
-
-        ``processed_images`` (from ``run``) and ``judge_scores`` (from ``judge``)
-        are independent: an image can be judged without ever being tagged. The
-        webapp's thumbnail/original/open-in-photos endpoints use the HTTP value
-        only as a lookup key and then read the DB-stored path, so resolving via
-        either table lets a judged-but-untagged image still preview (previously
-        only ``processed_images`` was consulted, so the Judge grid fell back to
-        filename text for every untagged image).
-        """
-        row = self._conn.execute(
-            "SELECT file_path FROM processed_images WHERE file_path = ? "
-            "UNION SELECT file_path FROM judge_scores WHERE file_path = ? "
-            "LIMIT 1",
-            (file_path, file_path),
-        ).fetchone()
-        return row[0] if row else None
-
-    def update_image_tags(self, file_path: str, tags: list[str]) -> None:
-        """Overwrite the tags list for an image."""
-        self._conn.execute(
-            "UPDATE processed_images SET tags = ? WHERE file_path = ?",
-            (json.dumps(tags), file_path),
-        )
-        self._conn.commit()
-
-    def update_image_cleanup(self, file_path: str, cleanup_class: str | None) -> None:
-        """Set or clear the cleanup_class for an image."""
-        self._conn.execute(
-            "UPDATE processed_images SET cleanup_class = ? WHERE file_path = ?",
-            (cleanup_class, file_path),
-        )
-        self._conn.commit()
-
-    def delete_image(self, file_path: str) -> bool:
-        """Delete a single row from ``processed_images`` by file_path.
-
-        Used by the Edit page after successfully removing the photo from
-        Apple Photos so a re-scan does not re-process the now-trashed
-        image. Returns True if a row was deleted, False if the path was
-        not present.
-        """
-        cur = self._conn.execute(
-            "DELETE FROM processed_images WHERE file_path = ?",
-            (file_path,),
-        )
-        self._conn.commit()
-        return cur.rowcount > 0
-
-    def iter_image_paths(self, batch_size: int = _DEFAULT_PATH_BATCH_SIZE) -> "Iterator[str]":
-        """Yield every ``file_path`` from ``processed_images`` in batches.
-
-        The drift-cleanup walk runs over a 22 k-row DB on the user's
-        machine; pulling everything into a single Python list pays an
-        unnecessary memory cost. A keyset cursor (``WHERE file_path >
-        last_seen``) is used instead of ``LIMIT … OFFSET`` so each page
-        costs O(log N) rather than a full or partial table scan, and rows
-        cannot be skipped or repeated if concurrent deletes occur between
-        pages (WAL mode, ``check_same_thread=False``).
-        """
-        last_path: str | None = None
-        while True:
-            if last_path is None:
-                rows = self._conn.execute(
-                    "SELECT file_path FROM processed_images ORDER BY file_path LIMIT ?",
-                    (int(batch_size),),
-                ).fetchall()
-            else:
-                rows = self._conn.execute(
-                    "SELECT file_path FROM processed_images"
-                    " WHERE file_path > ? ORDER BY file_path LIMIT ?",
-                    (last_path, int(batch_size)),
-                ).fetchall()
-            if not rows:
-                return
-            for (path,) in rows:
-                yield path
-            if len(rows) < batch_size:
-                return
-            last_path = rows[-1][0]
-
-    def delete_image_rows(self, paths: list[str]) -> int:
-        """Bulk-delete rows from ``processed_images`` matching *paths*.
-
-        Uses a single ``DELETE … WHERE file_path IN (…)`` statement so the
-        returned ``rowcount`` is accurate without any before/after COUNT(*)
-        queries. Returns the number of rows actually removed (which may be
-        smaller than ``len(paths)`` when some paths were already gone).
-        """
-        if not paths:
-            return 0
-        placeholders = ",".join("?" * len(paths))
-        cur = self._conn.execute(
-            f"DELETE FROM processed_images WHERE file_path IN ({placeholders})",  # nosec B608
-            paths,
-        )
-        self._conn.commit()
-        return cur.rowcount
-
-    @staticmethod
-    def _image_row_to_dict(row: tuple) -> dict:
-        """Convert a processed_images SELECT row to a metadata dict.
-
-        ``tags`` is returned as a parsed list of strings; iterating it as
-        ``for t in row['tags']`` yields tag values, not characters. The
-        legacy ``tags_list`` alias is kept for any callers that already
-        depend on it.
-        """
-        file_path: str = row[0]
-        tags_raw: str | None = row[1]
-        try:
-            tags_list: list[str] = json.loads(tags_raw) if tags_raw else []
-        except (json.JSONDecodeError, TypeError):
-            tags_list = []
-        weighted_raw = row[11] if len(row) > 11 else None
-        return {
-            "file_path": file_path,
-            "file_name": Path(file_path).name,
-            "tags": tags_list,
-            "tags_list": tags_list,
-            "scene_summary": row[2],
-            "processed_at": row[3],
-            "status": row[4],
-            "cleanup_class": row[5],
-            "scene_category": row[6],
-            "emotional_tone": row[7],
-            "event_hint": row[8],
-            "significance": row[9],
-            "error_message": row[10] if len(row) > 10 else None,
-            # Pulled from the optional LEFT JOIN with judge_scores. These
-            # are ``None`` for any image that has not been judged yet.
-            "judge_score": int(round(float(weighted_raw))) if weighted_raw is not None else None,
-            "judge_verdict": row[12] if len(row) > 12 else None,
-            "judge_reason": row[13] if len(row) > 13 else None,
-        }
-
-    # --- face pipeline methods ---
-
-    @staticmethod
-    def _embedding_to_blob(embedding: np.ndarray) -> bytes:
-        """Pack a 128-d float64 numpy array into a compact bytes blob."""
-        return struct.pack(f"{len(embedding)}d", *embedding.tolist())
-
-    @staticmethod
-    def _blob_to_embedding(blob: bytes) -> np.ndarray:
-        """Unpack a blob back into a numpy float64 array."""
-        import numpy as np
-
-        count = len(blob) // struct.calcsize("d")
-        return np.array(struct.unpack(f"{count}d", blob), dtype=np.float64)
-
-    def insert_face(
-        self,
-        image_path: str,
-        detection: FaceDetection,
-        embedding: np.ndarray | None = None,
-    ) -> int:
-        """Insert a detected face.
-
-        Args:
-            image_path: Path to the source image.
-            detection: Bounding box and confidence for the face.
-            embedding: Optional face encoding — a 1-D float64 numpy array
-                (128-d in practice) stored as a packed float64 blob.
-
-        Returns:
-            The new face row id.
-        """
-        blob = self._embedding_to_blob(embedding) if embedding is not None else None
-        cur = self._conn.execute(
-            """INSERT INTO faces (image_path, bbox_x, bbox_y, bbox_w, bbox_h, confidence, embedding)
-               VALUES (?, ?, ?, ?, ?, ?, ?)""",
-            (
-                image_path,
-                detection.bbox_x,
-                detection.bbox_y,
-                detection.bbox_w,
-                detection.bbox_h,
-                detection.confidence,
-                blob,
-            ),
-        )
-        self._conn.commit()
-        if cur.lastrowid is None:
-            raise RuntimeError("INSERT into faces did not return a row id")
-        return cur.lastrowid
-
-    def get_faces_by_uuid(self, uuid: str) -> list[dict]:
-        """Return all face rows whose image path stem matches the given UUID.
-
-        Handles both full paths (``/library/abc123.jpg``) and bare filenames
-        (``abc123.jpg``) by using two LIKE patterns.
-        """
-        rows = self._conn.execute(
-            "SELECT id, image_path, bbox_x, bbox_y, bbox_w, bbox_h, confidence, person_id "
-            "FROM faces WHERE image_path LIKE ? OR image_path LIKE ?",
-            (f"%/{uuid}.%", f"{uuid}.%"),
-        ).fetchall()
-        return [
-            {
-                "id": r[0],
-                "image_path": r[1],
-                "bbox_x": r[2],
-                "bbox_y": r[3],
-                "bbox_w": r[4],
-                "bbox_h": r[5],
-                "confidence": r[6],
-                "person_id": r[7],
-            }
-            for r in rows
-        ]
-
-    def has_photos_person(self, label: str) -> bool:
-        """Return True if a person with this label was already imported from Photos."""
-        return (
-            self._conn.execute(
-                "SELECT 1 FROM persons WHERE label = ? AND source = 'photos'",
-                (label,),
-            ).fetchone()
-            is not None
-        )
-
-    def get_photos_person_id(self, label: str) -> int | None:
-        """Return the person_id for an already-imported Photos.app person, or None."""
-        row = self._conn.execute(
-            "SELECT id FROM persons WHERE label = ? AND source = 'photos'",
-            (label,),
-        ).fetchone()
-        return row[0] if row else None
-
-    def mark_face_scanned(self, image_path: str) -> None:
-        """Record that an image has been fully face-scanned (even if 0 faces found)."""
-        self._conn.execute(
-            "INSERT OR IGNORE INTO face_scanned_images (image_path) VALUES (?)",
-            (image_path,),
-        )
-        self._conn.commit()
-
-    def is_face_scanned(self, image_path: str) -> bool:
-        """Return True if the image has already been face-scanned."""
-        return bool(
-            self._conn.execute(
-                "SELECT 1 FROM face_scanned_images WHERE image_path = ?", (image_path,)
-            ).fetchone()
-        )
-
-    def get_faces_for_image(self, image_path: str) -> list[dict]:
-        """Return all face rows for an image path."""
-        rows = self._conn.execute(
-            "SELECT id, bbox_x, bbox_y, bbox_w, bbox_h, confidence, person_id "
-            "FROM faces WHERE image_path = ?",
-            (image_path,),
-        ).fetchall()
-        return [
-            {
-                "id": r[0],
-                "bbox_x": r[1],
-                "bbox_y": r[2],
-                "bbox_w": r[3],
-                "bbox_h": r[4],
-                "confidence": r[5],
-                "person_id": r[6],
-            }
-            for r in rows
-        ]
-
-    def get_all_embeddings(self) -> list[tuple[int, np.ndarray]]:
-        """Return (face_id, embedding) for all faces that have embeddings."""
-        rows = self._conn.execute(
-            "SELECT id, embedding FROM faces WHERE embedding IS NOT NULL"
-        ).fetchall()
-        return [(r[0], self._blob_to_embedding(r[1])) for r in rows]
-
-    def get_clusterable_embeddings(self) -> list[tuple[int, np.ndarray]]:
-        """Return (face_id, embedding) for faces eligible for auto-clustering.
-
-        Only unassigned, non-ignored faces are returned. Faces already assigned
-        to a person (notably trusted / Photos-imported or manually confirmed
-        people) and trashed faces are excluded, so clustering never steals a
-        face away from a named person nor pulls a dismissed face back into a
-        cluster. ``clear_auto_persons`` releases auto-cluster faces back to the
-        unassigned pool before a recluster, so this set is exactly the faces
-        that should be (re)grouped.
-        """
-        rows = self._conn.execute(
-            "SELECT id, embedding FROM faces "
-            "WHERE embedding IS NOT NULL AND person_id IS NULL AND ignored = 0"
-        ).fetchall()
-        return [(r[0], self._blob_to_embedding(r[1])) for r in rows]
-
-    def get_embeddings_for_faces(self, face_ids: list[int]) -> dict[int, np.ndarray]:
-        """Return ``{face_id: embedding}`` for the given ids that have one.
-
-        Ids without a stored embedding (or unknown ids) are simply absent from
-        the result, so callers can look up by membership.
-        """
-        if not face_ids:
-            return {}
-        placeholders = ",".join("?" * len(face_ids))
-        rows = self._conn.execute(
-            f"SELECT id, embedding FROM faces WHERE id IN ({placeholders}) "  # nosec B608
-            "AND embedding IS NOT NULL",
-            list(face_ids),
-        ).fetchall()
-        return {r[0]: self._blob_to_embedding(r[1]) for r in rows}
-
-    def get_person_embeddings(self, person_id: int) -> list[np.ndarray]:
-        """Return the embeddings of the faces currently assigned to a person."""
-        rows = self._conn.execute(
-            "SELECT embedding FROM faces WHERE person_id = ? AND embedding IS NOT NULL",
-            (person_id,),
-        ).fetchall()
-        return [self._blob_to_embedding(r[0]) for r in rows]
-
-    def set_person_id(self, face_id: int, person_id: int) -> None:
-        """Assign a face to a person cluster."""
-        self._conn.execute("UPDATE faces SET person_id = ? WHERE id = ?", (person_id, face_id))
-        self._conn.commit()
-
-    def create_person(
-        self,
-        label: str = "",
-        confirmed: bool = False,
-        source: str = "auto",
-        trusted: bool = False,
-    ) -> int:
-        """Create a new person entry. Returns the person id."""
-        cur = self._conn.execute(
-            "INSERT INTO persons (label, confirmed, source, trusted) VALUES (?, ?, ?, ?)",
-            (label, 1 if confirmed else 0, source, 1 if trusted else 0),
-        )
-        self._conn.commit()
-        if cur.lastrowid is None:
-            raise RuntimeError("INSERT into persons did not return a row id")
-        return cur.lastrowid
-
-    def get_persons(self) -> list[PersonCluster]:
-        """Return all persons with their assigned face ids."""
-        persons = self._conn.execute(
-            "SELECT id, label, confirmed, source, trusted FROM persons"
-        ).fetchall()
-        # Fetch all face->person mappings in a single round-trip.
-        face_rows = self._conn.execute(
-            "SELECT person_id, id FROM faces WHERE person_id IS NOT NULL"
-        ).fetchall()
-        faces_by_person: dict[int, list[int]] = defaultdict(list)
-        for person_id, face_id in face_rows:
-            faces_by_person[person_id].append(face_id)
-        result = []
-        for pid, label, confirmed, source, trusted in persons:
-            result.append(
-                PersonCluster(
-                    person_id=pid,
-                    label=label,
-                    confirmed=bool(confirmed),
-                    face_ids=faces_by_person[pid],
-                    source=source or "auto",
-                    trusted=bool(trusted),
-                )
-            )
-        return result
-
-    def update_person_label(self, person_id: int, label: str) -> None:
-        """Update the display label for a person.
-
-        A non-empty label is treated as a manual confirmation — the person is
-        also marked ``confirmed=1, trusted=1`` so it survives re-clustering.
-        """
-        if label:
-            self._conn.execute(
-                "UPDATE persons SET label = ?, confirmed = 1, trusted = 1 WHERE id = ?",
-                (label, person_id),
-            )
-        else:
-            self._conn.execute("UPDATE persons SET label = ? WHERE id = ?", (label, person_id))
-        self._conn.commit()
-
-    def confirm_person(self, person_id: int) -> None:
-        """Mark a person cluster as confirmed and trusted.
-
-        Confirmed persons survive ``clear_auto_persons`` re-clustering and
-        their badge changes from AUTO → TRUSTED in the UI.
-        """
-        self._conn.execute(
-            "UPDATE persons SET confirmed = 1, trusted = 1 WHERE id = ?", (person_id,)
-        )
-        self._conn.commit()
-
-    def merge_persons(self, source_id: int, target_id: int) -> None:
-        """Reassign all faces from source_id to target_id, then delete source_id.
-
-        A non-existent ``source_id`` is a no-op (the UPDATE/DELETE simply affect
-        zero rows).
-
-        Raises:
-            ValueError: If ``target_id`` does not exist.
-        """
-        # Merging a person into itself would reassign its faces to the same id
-        # (a no-op) and then delete that very person — orphaning every face.
-        # Treat it as a no-op instead.
-        if source_id == target_id:
-            return
-        if not self._conn.execute("SELECT 1 FROM persons WHERE id = ?", (target_id,)).fetchone():
-            raise ValueError(f"merge target person {target_id} does not exist")
-        self._conn.execute(
-            "UPDATE faces SET person_id = ? WHERE person_id = ?", (target_id, source_id)
-        )
-        self._conn.execute("DELETE FROM persons WHERE id = ?", (source_id,))
-        self._conn.commit()
-
-    def delete_person(self, person_id: int) -> None:
-        """Delete a person and set person_id to NULL on all their faces."""
-        self._conn.execute("UPDATE faces SET person_id = NULL WHERE person_id = ?", (person_id,))
-        self._conn.execute("DELETE FROM persons WHERE id = ?", (person_id,))
-        self._conn.commit()
-
-    def confirm_persons(self, person_ids: list[int]) -> int:
-        """Mark several person clusters as confirmed and trusted in one transaction.
-
-        Args:
-            person_ids: Person ids to confirm. An empty list is a no-op.
-
-        Returns:
-            The number of person rows updated.
-        """
-        if not person_ids:
-            return 0
-        placeholders = ",".join("?" * len(person_ids))
-        cur = self._conn.execute(
-            f"UPDATE persons SET confirmed = 1, trusted = 1 WHERE id IN ({placeholders})",  # nosec B608
-            person_ids,
-        )
-        self._conn.commit()
-        return cur.rowcount
-
-    def delete_persons(self, person_ids: list[int]) -> int:
-        """Delete several persons and unassign their faces in one transaction.
-
-        Args:
-            person_ids: Person ids to delete. An empty list is a no-op.
-
-        Returns:
-            The number of person rows deleted.
-        """
-        if not person_ids:
-            return 0
-        placeholders = ",".join("?" * len(person_ids))
-        self._conn.execute(
-            f"UPDATE faces SET person_id = NULL WHERE person_id IN ({placeholders})",  # nosec B608
-            person_ids,
-        )
-        cur = self._conn.execute(
-            f"DELETE FROM persons WHERE id IN ({placeholders})",  # nosec B608
-            person_ids,
-        )
-        self._conn.commit()
-        return cur.rowcount
-
-    def clear_auto_persons(self) -> None:
-        """Delete all auto-clustered persons that are not trusted or confirmed.
-
-        Resets person_id to NULL on their faces so they can be re-clustered.
-        Persons imported from Photos (trusted=1) or manually confirmed are preserved.
-        """
-        auto_ids = [
-            r[0]
-            for r in self._conn.execute(
-                "SELECT id FROM persons WHERE trusted = 0 AND confirmed = 0"
-            ).fetchall()
-        ]
-        if not auto_ids:
-            return
-        placeholders = ",".join("?" * len(auto_ids))
-        self._conn.execute(
-            f"UPDATE faces SET person_id = NULL WHERE person_id IN ({placeholders})",  # nosec B608
-            auto_ids,
-        )
-        self._conn.execute(f"DELETE FROM persons WHERE id IN ({placeholders})", auto_ids)  # nosec B608
-        self._conn.commit()
-
-    def reset_all_faces(self, *, dry_run: bool = False) -> dict[str, int]:
-        """Delete every face, every person, and the face-scan cache.
-
-        The most destructive faces reset: a subsequent ``faces scan`` starts
-        from zero, re-detecting and re-clustering everything. Trusted and
-        Photos-imported persons are **not** spared. Image tagging/geocoding
-        progress in ``processed_images`` is untouched.
-
-        Args:
-            dry_run: When True, only count what would be removed; delete nothing.
-
-        Returns:
-            ``{"faces": n, "persons": n, "scanned_images": n}`` — rows removed
-            (or that would be removed, for ``dry_run``).
-        """
-        counts = {
-            "faces": self._conn.execute("SELECT COUNT(*) FROM faces").fetchone()[0],
-            "persons": self._conn.execute("SELECT COUNT(*) FROM persons").fetchone()[0],
-            "scanned_images": self._conn.execute(
-                "SELECT COUNT(*) FROM face_scanned_images"
-            ).fetchone()[0],
-        }
-        if dry_run:
-            return counts
-        self._conn.execute("DELETE FROM faces")
-        self._conn.execute("DELETE FROM persons")
-        self._conn.execute("DELETE FROM face_scanned_images")
-        self._conn.commit()
-        return counts
-
-    def reset_untrusted_faces(self, *, dry_run: bool = False) -> dict[str, int]:
-        """Delete non-trusted faces and clusters, preserving trusted people.
-
-        Removes every person that is neither trusted nor confirmed, and every
-        non-ignored face not assigned to a surviving trusted person, then
-        prunes the face-scan cache for images that have no face left so the
-        next scan re-detects them. Kept: trusted / Photos-imported persons and
-        the faces assigned to them, AND ignored ("trash") faces — those are
-        explicit user curation, so they survive here; use ``reset_all_faces``
-        to clear them too. (An image that still holds a trusted or trashed
-        face stays cached, so its faces are not re-detected — nor the kept
-        face duplicated — on the next scan.)
-
-        Args:
-            dry_run: When True, only count what would be removed; delete nothing.
-
-        Returns:
-            ``{"faces": n, "persons": n, "scanned_images": n}``.
-        """
-        # All SQL below is constant (no interpolated values). A face is deleted
-        # only when it is NOT ignored AND not assigned to a trusted/confirmed
-        # person. The scan-cache survivor set is images that still have a
-        # trusted OR an ignored face after the delete.
-        counts = {
-            "faces": self._conn.execute(
-                "SELECT COUNT(*) FROM faces WHERE ignored = 0 AND (person_id IS NULL "
-                "OR person_id NOT IN (SELECT id FROM persons WHERE trusted = 1 OR confirmed = 1))"
-            ).fetchone()[0],
-            "persons": self._conn.execute(
-                "SELECT COUNT(*) FROM persons WHERE trusted = 0 AND confirmed = 0 "
-                "AND id NOT IN (SELECT person_id FROM faces "
-                "WHERE ignored = 1 AND person_id IS NOT NULL)"
-            ).fetchone()[0],
-            "scanned_images": self._conn.execute(
-                "SELECT COUNT(*) FROM face_scanned_images WHERE image_path NOT IN "
-                "(SELECT DISTINCT image_path FROM faces WHERE ignored = 1 OR person_id IN "
-                "(SELECT id FROM persons WHERE trusted = 1 OR confirmed = 1))"
-            ).fetchone()[0],
-        }
-        if dry_run:
-            return counts
-        self._conn.execute(
-            "DELETE FROM faces WHERE ignored = 0 AND (person_id IS NULL "
-            "OR person_id NOT IN (SELECT id FROM persons WHERE trusted = 1 OR confirmed = 1))"
-        )
-        # Keep any untrusted person that still owns a surviving (ignored/trash)
-        # face — deleting it would leave that face pointing at a non-existent
-        # person (SQLite FK enforcement is off, so the dangling row is silent).
-        # ``ignored = 1`` faces are never touched by the face delete above, so
-        # this predicate matches the count computed before the deletes.
-        self._conn.execute(
-            "DELETE FROM persons WHERE trusted = 0 AND confirmed = 0 "
-            "AND id NOT IN (SELECT person_id FROM faces "
-            "WHERE ignored = 1 AND person_id IS NOT NULL)"
-        )
-        # Only trusted and ignored faces survive now, so "no face left" is just
-        # "no face row left" for the image.
-        self._conn.execute(
-            "DELETE FROM face_scanned_images WHERE image_path NOT IN "
-            "(SELECT DISTINCT image_path FROM faces)"
-        )
-        self._conn.commit()
-        return counts
-
-    def count_auto_persons(self) -> int:
-        """Return the number of auto-clustered persons (trusted=0 AND confirmed=0)."""
-        return self._conn.execute(
-            "SELECT COUNT(*) FROM persons WHERE trusted = 0 AND confirmed = 0"
-        ).fetchone()[0]
-
-    def get_auto_person_ids(self) -> set[int]:
-        """Return the ids of auto-clustered persons (trusted=0 AND confirmed=0).
-
-        These are the only assignments a UUID-authoritative re-link (Photos
-        import) may reclaim a face from; trusted/confirmed assignments are
-        never disturbed.
-        """
-        return {
-            r[0]
-            for r in self._conn.execute(
-                "SELECT id FROM persons WHERE trusted = 0 AND confirmed = 0"
-            ).fetchall()
-        }
-
-    def ignore_face(self, face_id: int) -> None:
-        """Mark a face as ignored so it is excluded from auto-clustering."""
-        self._conn.execute(
-            "UPDATE faces SET ignored = 1, person_id = NULL WHERE id = ?", (face_id,)
-        )
-        self._conn.commit()
-
-    def restore_face(self, face_id: int) -> None:
-        """Remove the ignored flag from a face, returning it to the unassigned pool."""
-        self._conn.execute("UPDATE faces SET ignored = 0 WHERE id = ?", (face_id,))
-        self._conn.commit()
-
-    def get_ignored_faces(self) -> list[dict]:
-        """Return all faces marked as ignored (the trash bin)."""
-        rows = self._conn.execute(
-            "SELECT id, image_path, bbox_x, bbox_y, bbox_w, bbox_h, confidence "
-            "FROM faces WHERE ignored = 1"
-        ).fetchall()
-        return [
-            {
-                "id": r[0],
-                "image_path": r[1],
-                "bbox_x": r[2],
-                "bbox_y": r[3],
-                "bbox_w": r[4],
-                "bbox_h": r[5],
-                "confidence": r[6],
-            }
-            for r in rows
-        ]
-
-    def get_unassigned_faces(self) -> list[dict]:
-        """Return faces that have no person assignment and are not ignored."""
-        rows = self._conn.execute(
-            "SELECT id, image_path, bbox_x, bbox_y, bbox_w, bbox_h, confidence "
-            "FROM faces WHERE person_id IS NULL AND ignored = 0"
-        ).fetchall()
-        return [
-            {
-                "id": r[0],
-                "image_path": r[1],
-                "bbox_x": r[2],
-                "bbox_y": r[3],
-                "bbox_w": r[4],
-                "bbox_h": r[5],
-                "confidence": r[6],
-            }
-            for r in rows
-        ]
-
-    def unassign_face(self, face_id: int) -> None:
-        """Remove the person assignment from a face."""
-        self._conn.execute("UPDATE faces SET person_id = NULL WHERE id = ?", (face_id,))
-        self._conn.commit()
-
-    def get_faces_for_person(self, person_id: int) -> list[dict]:
-        """Return all faces assigned to a person."""
-        rows = self._conn.execute(
-            "SELECT id, image_path, bbox_x, bbox_y, bbox_w, bbox_h, confidence "
-            "FROM faces WHERE person_id = ?",
-            (person_id,),
-        ).fetchall()
-        return [
-            {
-                "id": r[0],
-                "image_path": r[1],
-                "bbox_x": r[2],
-                "bbox_y": r[3],
-                "bbox_w": r[4],
-                "bbox_h": r[5],
-                "confidence": r[6],
-            }
-            for r in rows
-        ]
-
-    def get_assigned_faces(self) -> list[dict]:
-        """Return all faces that have a person assignment."""
-        rows = self._conn.execute(
-            "SELECT id, image_path FROM faces WHERE person_id IS NOT NULL"
-        ).fetchall()
-        return [{"id": r[0], "image_path": r[1]} for r in rows]
-
-    def get_face_count(self) -> int:
-        """Return total number of detected faces."""
-        return self._conn.execute("SELECT COUNT(*) FROM faces").fetchone()[0]
-
-    def get_face_by_id(self, face_id: int) -> dict | None:
-        """Return a single face record by id, or None if not found."""
-        row = self._conn.execute(
-            "SELECT id, image_path, bbox_x, bbox_y, bbox_w, bbox_h, confidence "
-            "FROM faces WHERE id = ?",
-            (face_id,),
-        ).fetchone()
-        if row is None:
-            return None
-        return {
-            "id": row[0],
-            "image_path": row[1],
-            "bbox_x": row[2],
-            "bbox_y": row[3],
-            "bbox_w": row[4],
-            "bbox_h": row[5],
-            "confidence": row[6],
-        }
-
-    def save_judge_result(self, result: "JudgeResult") -> None:
-        """Persist a judge scoring result. Replaces any existing entry for the same file."""
-        s = result.scores
-        self._conn.execute(
-            """
-            INSERT OR REPLACE INTO judge_scores
-                (file_path, scored_at, weighted_score, core_score, visible_score, verdict,
-                 impact, story_subject, composition_center, lighting, creativity_style,
-                 color_mood, presentation_crop, technical_excellence, focus_sharpness,
-                 exposure_tonal, noise_cleanliness, subject_separation, edit_integrity,
-                 reason)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-            (
-                result.file_path,
-                datetime.now(timezone.utc).isoformat(),
-                result.weighted_score,
-                result.core_score,
-                result.visible_score,
-                s.verdict,
-                s.impact,
-                s.story_subject,
-                s.composition_center,
-                s.lighting,
-                s.creativity_style,
-                s.color_mood,
-                s.presentation_crop,
-                s.technical_excellence,
-                s.focus_sharpness,
-                s.exposure_tonal,
-                s.noise_cleanliness,
-                s.subject_separation,
-                s.edit_integrity,
-                s.reason,
-            ),
-        )
-        self._conn.commit()
-
-    def get_judge_result(self, file_path: str) -> dict | None:
-        """Return judge scores for a file, or None if not found.
-
-        Scores are integers in 1-10. The underlying ``judge_scores`` table
-        uses REAL columns for backward compatibility, so values are cast to
-        ``int`` here (rounded for any legacy float rows from the old 1-5
-        scale).
-        """
-        row = self._conn.execute(
-            """SELECT weighted_score, core_score, visible_score, verdict,
-                      impact, story_subject, composition_center, lighting,
-                      creativity_style, color_mood, presentation_crop,
-                      technical_excellence, focus_sharpness, exposure_tonal,
-                      noise_cleanliness, subject_separation, edit_integrity,
-                      scored_at, reason
-               FROM judge_scores WHERE file_path = ?""",
-            (file_path,),
-        ).fetchone()
-        if row is None:
-            return None
-
-        def _i(v: Any) -> int:
-            return int(round(float(v))) if v is not None else 0
-
-        return {
-            "file_path": file_path,
-            "weighted_score": _i(row[0]),
-            "core_score": _i(row[1]),
-            "visible_score": _i(row[2]),
-            "verdict": row[3],
-            "scored_at": row[17],
-            "reason": row[18] if len(row) > 18 else None,
-            "scores": {
-                "impact": _i(row[4]),
-                "story_subject": _i(row[5]),
-                "composition_center": _i(row[6]),
-                "lighting": _i(row[7]),
-                "creativity_style": _i(row[8]),
-                "color_mood": _i(row[9]),
-                "presentation_crop": _i(row[10]),
-                "technical_excellence": _i(row[11]),
-                "focus_sharpness": _i(row[12]),
-                "exposure_tonal": _i(row[13]),
-                "noise_cleanliness": _i(row[14]),
-                "subject_separation": _i(row[15]),
-                "edit_integrity": _i(row[16]),
-            },
-        }
-
-    def get_all_judge_results(self, limit: int | None = _DEFAULT_JUDGE_RESULTS_LIMIT) -> list[dict]:
-        """Return all judge scores ordered by weighted_score descending.
-
-        Args:
-            limit: Max rows to return. None = no limit (caution: may be large).
-
-        Returns:
-            List of dicts with keys: file_path, file_name, weighted_score,
-            core_score, visible_score, verdict, scored_at.
-        """
-        limit_clause = f"LIMIT {int(limit)}" if limit is not None else ""
-        rows = self._conn.execute(
-            "SELECT file_path, weighted_score, core_score, visible_score, verdict, "
-            "scored_at, reason "
-            "FROM judge_scores "
-            "ORDER BY weighted_score DESC " + limit_clause  # nosec B608
-        ).fetchall()
-
-        def _i(v: Any) -> int:
-            return int(round(float(v))) if v is not None else 0
-
-        return [
-            {
-                "file_path": row[0],
-                "file_name": Path(row[0]).name,
-                "weighted_score": _i(row[1]),
-                "core_score": _i(row[2]),
-                "visible_score": _i(row[3]),
-                "verdict": row[4],
-                "scored_at": row[5],
-                "reason": row[6] if len(row) > 6 else None,
-            }
-            for row in rows
-        ]
-
-    # Whitelisted ORDER BY clauses for query_judge_results. Keeping the
-    # rating_* keys distinct from the legacy ``judge_*`` clauses on
-    # ``query_images`` keeps the Judge UI's API surface independent.
-    _JUDGE_QUERY_SORTS: dict[str, str] = {
-        "rating_desc": "js.weighted_score DESC NULLS LAST, pi.file_path ASC",
-        "rating_asc": "js.weighted_score ASC NULLS LAST, pi.file_path ASC",
-        "path_asc": "pi.file_path ASC",
-        "path_desc": "pi.file_path DESC",
-        "shot_desc": "pi.image_date DESC NULLS LAST, pi.file_path ASC",
-        "shot_asc": "pi.image_date ASC NULLS LAST, pi.file_path ASC",
-    }
-
-    def query_judge_results(
-        self,
-        offset: int = 0,
-        limit: int = 50,
-        sort: str = "rating_desc",
-        min_rating: int | None = None,
-        max_rating: int | None = None,
-    ) -> dict:
-        """Return paginated judged images joined with their image metadata.
-
-        The Judge page wants every record that has a ``judge_scores`` row
-        plus the matching ``processed_images`` columns it renders next to
-        the rating (file_name, scene_summary, image_date, location,
-        cleanup_class). Caller-supplied rating bounds are clamped to
-        ``[1, 10]`` rather than rejected, so the JS can pass the raw input
-        without pre-validating.
-
-        Args:
-            offset: Row offset for pagination.
-            limit: Max rows to return.
-            sort: One of ``rating_desc`` (default), ``rating_asc``,
-                ``path_asc``, ``path_desc``, ``shot_desc``, ``shot_asc``.
-            min_rating: Inclusive lower bound on weighted_score (1-10).
-            max_rating: Inclusive upper bound on weighted_score (1-10).
-
-        Returns:
-            ``{"items": [...], "total": <int>}``. Each item carries the
-            keys consumed by the Judge UI: file_path, file_name,
-            weighted_score, reason, verdict, image_date, scene_summary,
-            nearest_city, nearest_country, cleanup_class.
-        """
-        conditions: list[str] = []
-        params: list[object] = []
-        if min_rating is not None:
-            clamped_min = max(1, min(10, int(min_rating)))
-            conditions.append("js.weighted_score >= ?")
-            params.append(clamped_min)
-        if max_rating is not None:
-            clamped_max = max(1, min(10, int(max_rating)))
-            conditions.append("js.weighted_score <= ?")
-            params.append(clamped_max)
-        where = ("WHERE " + " AND ".join(conditions)) if conditions else ""
-        order_clause = self._JUDGE_QUERY_SORTS.get(sort, self._JUDGE_QUERY_SORTS["rating_desc"])
-
-        count_query = (  # nosec B608
-            "SELECT COUNT(*) FROM judge_scores js "
-            "LEFT JOIN processed_images pi ON pi.file_path = js.file_path " + where  # nosec B608
-        )
-        total = self._conn.execute(count_query, params).fetchone()[0]
-
-        list_query = (  # nosec B608
-            "SELECT js.file_path, js.weighted_score, js.reason, js.verdict, "
-            "pi.scene_summary, pi.nearest_city, pi.nearest_country, "
-            "pi.cleanup_class, pi.image_date, js.scored_at "
-            "FROM judge_scores js "
-            "LEFT JOIN processed_images pi ON pi.file_path = js.file_path "
-            + where  # nosec B608
-            + f" ORDER BY {order_clause} LIMIT ? OFFSET ?"  # nosec B608
-        )
-        params_list = [*params, int(limit), int(offset)]
-        rows = self._conn.execute(list_query, params_list).fetchall()
-
-        def _i(v: Any) -> int | None:
-            return int(round(float(v))) if v is not None else None
-
-        items = [
-            {
-                "file_path": row[0],
-                "file_name": Path(row[0]).name,
-                "weighted_score": _i(row[1]),
-                "reason": row[2],
-                "verdict": row[3],
-                "scene_summary": row[4],
-                "nearest_city": row[5],
-                "nearest_country": row[6],
-                "cleanup_class": row[7],
-                "image_date": row[8],
-                "scored_at": row[9],
-            }
-            for row in rows
-        ]
-        return {"items": items, "total": int(total)}
-
-    def is_fresh(self, file_path: Path) -> bool:
-        """Return True if DB has a row for this file and size/mtime still match."""
-        row = self._conn.execute(
-            "SELECT file_size, file_mtime FROM processed_images WHERE file_path = ?",
-            (str(file_path),),
-        ).fetchone()
-        if row is None:
-            return False
-        try:
-            stat = file_path.stat()
-        except OSError:
-            return False
-        return row[0] == stat.st_size and row[1] == stat.st_mtime
-
-    def get_cached_result(self, file_path: Path) -> ImageResult | None:
-        """Return an ImageResult built from the stored DB row, or None if not found."""
-        row = self._conn.execute(
-            """
-            SELECT tags, scene_summary, status, error_message,
-                   scene_category, emotional_tone, cleanup_class,
-                   has_text, text_summary, event_hint, significance
-            FROM processed_images
-            WHERE file_path = ?
-            """,
-            (str(file_path),),
-        ).fetchone()
-        if row is None:
-            return None
-        (
-            tags_raw,
-            scene_summary,
-            status,
-            error_message,
-            scene_category,
-            emotional_tone,
-            cleanup_class,
-            has_text_int,
-            text_summary,
-            event_hint,
-            significance,
-        ) = row
-        try:
-            tags: list[str] = json.loads(tags_raw) if tags_raw else []
-        except (json.JSONDecodeError, TypeError):
-            tags = []
-        return ImageResult(
-            file_path=str(file_path),
-            file_name=file_path.name,
-            tags=tags,
-            scene_summary=scene_summary,
-            processing_status=status if status is not None else "ok",
-            error_message=error_message,
-            scene_category=scene_category,
-            emotional_tone=emotional_tone,
-            cleanup_class=cleanup_class,
-            has_text=bool(has_text_int),
-            text_summary=text_summary,
-            event_hint=event_hint,
-            significance=significance,
-        )
-
-    def is_complete_cached(self, file_path: Path) -> bool:
-        """Return True if the DB has a fresh, successful row with non-empty tags.
-
-        Combines the freshness (size+mtime), status, and non-empty-tags checks
-        into a single query so the skip decision costs one SELECT plus one
-        ``stat()`` — the fast path used by ``run --skip-existing`` to bypass
-        EXIF, geocoding, and write-back for already-processed photos.
-        """
-        row = self._conn.execute(
-            "SELECT file_size, file_mtime, status, tags FROM processed_images WHERE file_path = ?",
-            (str(file_path),),
-        ).fetchone()
-        if row is None:
-            return False
-        size, mtime, status, tags_raw = row
-        if status != "ok":
-            return False
-        try:
-            tags: list[str] = json.loads(tags_raw) if tags_raw else []
-        except (json.JSONDecodeError, TypeError):
-            tags = []
-        if not tags:
-            return False
-        try:
-            stat = file_path.stat()
-        except OSError:
-            return False
-        return size == stat.st_size and mtime == stat.st_mtime
-
-    def has_usable_model_result(self, file_path: Path) -> bool:
-        """Return True if the DB has a fresh row with non-empty tags.
-
-        A tags value that is not valid JSON is treated as no tags, so a row
-        with a corrupt tags blob is reported as not usable (returns False)
-        rather than raising.
-        """
-        if not self.is_fresh(file_path):
-            return False
-        row = self._conn.execute(
-            "SELECT tags FROM processed_images WHERE file_path = ?",
-            (str(file_path),),
-        ).fetchone()
-        if row is None:
-            return False
-        try:
-            tags: list[str] = json.loads(row[0]) if row[0] else []
-        except (json.JSONDecodeError, TypeError):
-            tags = []  # malformed JSON -> treat as no tags
-        return len(tags) > 0
-
-    def update_missing_fields(self, file_path: Path, result: ImageResult) -> None:
-        """Update only NULL columns from result.
-
-        ``has_text`` additionally treats 0 as unset (it can be upgraded
-        0 -> 1 but never downgraded). Other existing non-null values are
-        never overwritten.
-        """
-        self._conn.execute(
-            """
-            UPDATE processed_images SET
-                scene_summary  = COALESCE(scene_summary,  ?),
-                scene_category = COALESCE(scene_category, ?),
-                emotional_tone = COALESCE(emotional_tone, ?),
-                cleanup_class  = COALESCE(cleanup_class,  ?),
-                has_text       = CASE WHEN has_text = 0 OR has_text IS NULL
-                                      THEN ? ELSE has_text END,
-                text_summary   = COALESCE(text_summary,   ?),
-                event_hint     = COALESCE(event_hint,     ?),
-                significance   = COALESCE(significance,   ?)
-            WHERE file_path = ?
-            """,
-            (
-                result.scene_summary,
-                result.scene_category,
-                result.emotional_tone,
-                result.cleanup_class,
-                1 if result.has_text else 0,
-                result.text_summary,
-                result.event_hint,
-                result.significance,
-                str(file_path),
-            ),
-        )
-        self._conn.commit()
-
-    def __enter__(self) -> "ProgressDB":
-        """Enter the context manager, returning this instance."""
-        return self
-
-    def __exit__(self, exc_type: object, exc_val: object, exc_tb: object) -> None:
-        """Exit the context manager, closing the connection via :meth:`close`."""
-        self.close()
-
-    def close(self) -> None:
-        """Close the underlying SQLite connection.
-
-        The instance must not be used afterwards.
-        """
-        self._conn.close()
+__all__ = [
+    "_DEFAULT_JUDGE_RESULTS_LIMIT",
+    "_DEFAULT_PATH_BATCH_SIZE",
+    "FaceDetection",
+    "ImageResult",
+    "PersonCluster",
+    "ProgressDB",
+]
